Fix issue #1
If HTML was created using a here-string it presents as one big string,
where the script was expecting an array (1 line per element).  I simply
did a split on `n`r (carriage return/line feed) on whatever I get.  If
it's a here-string it will now be split into an array, if it isn't it
won't because it won't match the split criteria.